[feat] 소수 적용

+++ src/main/java/Calculator.java
@@ -1,17 +1,17 @@
 public class Calculator {
     public double getResult(String[] formula) {
-        double result = Integer.parseInt(formula[0]);
+        double result = Double.parseDouble(formula[0]);
 
         for(int i=1; i<formula.length; i+=2) {
             String op = formula[i];
             if(op.equals("+")) {
-                result += Integer.parseInt(formula[i+1]);
+                result += Double.parseDouble(formula[i+1]);
             } else if(op.equals("-")) {
-                result -= Integer.parseInt(formula[i+1]);
+                result -= Double.parseDouble(formula[i+1]);
             } else if(op.equals("*")) {
-                result *= Integer.parseInt(formula[i+1]);
+                result *= Double.parseDouble(formula[i+1]);
             } else if(op.equals("/")) {
-                result /= Integer.parseInt(formula[i+1]);
+                result /= Double.parseDouble(formula[i+1]);
             }
         }
         return result;
